requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
minio==7.2.9

SQLAlchemy==2.0.35
psycopg[binary]==3.2.10
python-dateutil==2.9.0.post0
